Move tests for happinessutil.py into test_happiness.py.
