+++ src/allmydata/test/common_util.py
@@ -5,14 +5,13 @@ from random import randrange
 from six.moves import StringIO
 
 from twisted.internet import reactor, defer
-from twisted.python import failure
 from twisted.trial import unittest
 
 from ..util.assertutil import precondition
 from allmydata.util.encodingutil import (unicode_platform, get_filesystem_encoding,
                                          get_io_encoding)
 from ..scripts import runner
-from .common_py3 import SignalMixin
+from .common_py3 import SignalMixin, ShouldFailMixin  # noqa
 
 
 def skip_if_cannot_represent_filename(u):
@@ -94,28 +93,6 @@ class StallMixin(object):
     def stall(self, res=None, delay=1):
         d = defer.Deferred()
         reactor.callLater(delay, d.callback, res)
-        return d
-
-class ShouldFailMixin(object):
-
-    def shouldFail(self, expected_failure, which, substring,
-                   callable, *args, **kwargs):
-        assert substring is None or isinstance(substring, str)
-        d = defer.maybeDeferred(callable, *args, **kwargs)
-        def done(res):
-            if isinstance(res, failure.Failure):
-                res.trap(expected_failure)
-                if substring:
-                    self.failUnless(substring in str(res),
-                                    "%s: substring '%s' not in '%s'"
-                                    % (which, substring, str(res)))
-                # return the Failure for further analysis, but in a form that
-                # doesn't make the Deferred chain think that we failed.
-                return [res]
-            else:
-                self.fail("%s was supposed to raise %s, not get '%s'" %
-                          (which, expected_failure, res))
-        d.addBoth(done)
         return d
